requirements.txt
# Smart Records System - Python Dependencies
# Install all dependencies using: pip install -r requirements.txt

# GUI Framework (Required)
customtkinter>=5.2.0

# PDF Report Generation (Required for PDF export feature)
reportlab>=3.6.0

# MySQL Database Support (Required if using MySQL database)
# Uncomment the line below if you plan to use MySQL:
mysql-connector-python>=8.0.0

# Note: The following are part of Python standard library and don't need to be installed:
# - tkinter (used for messagebox and menu - still needed)
# - hashlib (Password hashing)
# - datetime (Date handling)
# - re (Regular expressions)
# - os (File operations)
# - subprocess (For setup scripts)

